requirements.txt
PyYAML==3.13
selenium==3.14.1
urllib3==1.23
